requirements.txt
Django==4.0.5
django-environ==0.9.0
djangorestframework==3.13.1
requests==2.28.0